fibonacci numbers set point func is added.

+++ src/main/java/demo/entity/Issue.java
@@ -37,7 +37,7 @@ public class Issue {
 
 	private String description;
 	
-	private Double storyPoint;
+	private int storyPoint;
 	
 	private Boolean isRevealed;
